Restructured core packages
Restructured the packages in core to make them more hierarchical and
keep them managable as we add new systems and charts.
