+++ core/src/main/java/org/cirdles/topsoil/plot/upb/uncertainty/UncertaintyEllipsePlotProperties.java
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2016 CIRDLES.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.cirdles.topsoil.plot.upb.uncertainty;
+
+/**
+ * Created by johnzeringue on 11/9/15.
+ */
+public final class UncertaintyEllipsePlotProperties {
+
+    public static final String TITLE = "Title";
+    public static final String X_AXIS = "X Axis";
+    public static final String Y_AXIS = "Y Axis";
+    public static final String UNCERTAINTY = "Uncertainty";
+    public static final String ELLIPSE_FILL_COLOR = "Ellipse Fill Color";
+    public static final String LAMBDA_235 = "LAMBDA_235";
+    public static final String LAMBDA_238 = "LAMBDA_238";
+
+    private UncertaintyEllipsePlotProperties() {
+        // prevents this class from being instantiated
+    }
+
+}
