+++ core/src/main/resources/org/cirdles/topsoil/plot/scatter/ScatterPlot.js
