+++ core/src/test/java/org/cirdles/topsoil/plot/upb/uncertainty/UncertaintyEllipsePlotTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015 CIRDLES.
+ * Copyright 2016 CIRDLES.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.cirdles.topsoil.plot.standard;
+package org.cirdles.topsoil.plot.upb.uncertainty;
 
 import javafx.scene.Parent;
 import javafx.scene.Scene;
