+++ core/src/test/java/org/cirdles/topsoil/plot/uth/evolution/EvolutionPlotTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015 CIRDLES.
+ * Copyright 2016 CIRDLES.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,18 +13,32 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.cirdles.topsoil.plot.standard;
+package org.cirdles.topsoil.plot.uth.evolution;
+
+import org.cirdles.topsoil.plot.Plot;
+import org.junit.Before;
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.rules.Timeout;
 
 /**
- * Created by johnzeringue on 11/9/15.
+ * Created by johnzeringue on 1/31/16.
  */
-public final class ScatterPlotProperties {
+public class EvolutionPlotTest {
 
-    public static final String TITLE = "Title";
-    public static final String POINT_FILL_COLOR = "Point Fill Color";
+    @Rule
+    public Timeout timeout = Timeout.seconds(5);
 
-    private ScatterPlotProperties() {
-        // prevents this class from being instantiated
+    private Plot plot;
+
+    @Before
+    public void setUp() {
+        plot = new EvolutionPlot();
+    }
+
+    @Test
+    public void testInitialize() {
+        // no-op
     }
 
 }
